+++ src/api/controllers/create_report.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 
-from api.datatype.Eatery import EateryID
+from eatery.datatype.Eatery import EateryID
 from api.models import ReportStore
