+++ src/api/dfg/nodes/schedule/ClosedSchedule.py
@@ -1,4 +1,4 @@
-from api.datatype.Eatery import Eatery, EateryID
+from eatery.datatype.Eatery import Eatery, EateryID
 from api.dfg.nodes.DfgNode import DfgNode
 from api.models.EventScheduleModel import ExceptionType, ScheduleException
