+++ src/api/models/EventScheduleModel.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from api.models.EateryModel import EateryStore
+from eatery.models import EateryStore
 from api.models.MenuModel import MenuStore
 from django.core.validators import validate_comma_separated_integer_list
 from django.db import models
