+++ src/api/models/__init__.py
@@ -2,7 +2,6 @@
 # In this app, models should be imported directly from api.models, not from api.models.package
 
 from .AlertModel import AlertStore
-from .EateryModel import EateryStore
 from .EventScheduleModel import (
     EventSchedule,
     RepeatingEventSchedule,
