+++ src/api/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
-from api.views import MainDfgView, UpdateView, ReportView
+from api.views import MainDfgView, ReportView
+from eatery.views import UpdateView
 
 urlpatterns = [
     path("", MainDfgView.as_view(), name="main"),
