+++ src/eatery/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+import eatery.models as models
+
+# Register your models here.
+
+admin.site.register(models.EateryStore)
