+++ src/eatery/datatype/EateryAlert.py
