+++ src/eatery_blue_backend/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "api",
+    "eatery",
     "rest_framework",
 ]
